relatorios/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
oracledb==2.4.1
python-dotenv==1.0.1
